+++ esm6/client.js
@@ -1,4 +1,5 @@
 import { StompHandler } from "./stomp-handler";
+import { Versions } from "./versions";
 /**
  * STOMP Client Class.
  */
@@ -8,6 +9,16 @@ var Client = /** @class */ (function () {
      */
     function Client(conf) {
         if (conf === void 0) { conf = {}; }
+        /**
+         * STOMP versions to attempt during STOMP handshake. By default versions `1.0`, `1.1`, and `1.2` are attempted.
+         *
+         * Example:
+         * ```javascript
+         *        // Try only versions 1.0 and 1.1
+         *        client.stompVersions = new Versions(['1.0', '1.1'])
+         * ```
+         */
+        this.stompVersions = Versions.default;
         /**
          *  automatically reconnect with delay in milliseconds, set to 0 to disable.
          */
@@ -58,12 +69,12 @@ var Client = /** @class */ (function () {
         enumerable: true,
         configurable: true
     });
-    Object.defineProperty(Client.prototype, "version", {
+    Object.defineProperty(Client.prototype, "connectedVersion", {
         /**
          * version of STOMP protocol negotiated with the server, READONLY
          */
         get: function () {
-            return this._stompHandler ? this._stompHandler.version : undefined;
+            return this._stompHandler ? this._stompHandler.connectedVersion : undefined;
         },
         enumerable: true,
         configurable: true
@@ -112,6 +123,7 @@ var Client = /** @class */ (function () {
         this._webSocket = this._createWebSocket();
         this._stompHandler = new StompHandler(this, this._webSocket, {
             debug: this.debug,
+            stompVersions: this.stompVersions,
             connectHeaders: this.connectHeaders,
             disconnectHeaders: this.disconnectHeaders,
             heartbeatIncoming: this.heartbeatIncoming,
@@ -151,7 +163,13 @@ var Client = /** @class */ (function () {
         this._stompHandler.start();
     };
     Client.prototype._createWebSocket = function () {
-        var webSocket = this.webSocketFactory ? this.webSocketFactory() : new WebSocket(this.brokerURL);
+        var webSocket;
+        if (this.webSocketFactory) {
+            webSocket = this.webSocketFactory();
+        }
+        else {
+            webSocket = new WebSocket(this.brokerURL, this.stompVersions.protocolVersions());
+        }
         webSocket.binaryType = "arraybuffer";
         return webSocket;
     };