+++ esm6/compatibility/stomp.js
@@ -29,7 +29,7 @@ var Stomp = /** @class */ (function () {
         // instead.
         // See remarks on the function Stomp.over
         if (protocols == null) {
-            protocols = Versions.protocolVersions();
+            protocols = Versions.default.protocolVersions();
         }
         var ws_fn = function () {
             var klass = Stomp.WebSocketClass || WebSocket;